refactor[#193] : 서로 다른 도메인에서 사용 되는 닉네임 중복 검사 로직 shared 로 이동
