+++ src/shared/constants/index.ts
@@ -1,3 +1,5 @@
+const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
+
 export const colors = {
   grey: {
     0: "#FFFFFF",
@@ -115,3 +117,7 @@ export const ROUTER_PATH = {
   ACCOUNT: "/my-page/setting/manage-account",
   MODIFY_MY_INFO: "/my-page/setting/modify-userinfo",
 } as const;
+
+export const API_END_POINT = {
+  DUPLICATE_NICKNAME: `${API_BASE_URL}/users/nickname`,
+};
